+++ src/plugins/minimize/toggle.js
@@ -1,0 +1,9 @@
+import { Model } from '@converse/skeletor/src/model.js';
+
+const MinimizedChatsToggle = Model.extend({
+    defaults: {
+        'collapsed': false
+    }
+});
+
+export default MinimizedChatsToggle;
